Moved local modules a directory up
All modules in local modules were moved out of the folder and into the
main project. The previous model_inference folder was deleted in favor
of the rewritten model_inference folder that was in local_modules.

plot_utilites folder was renamed to plots

main now tries to read data via modbus